Link existing entities/events in graph if found, rather than duplicate
