+++ cdse_covid/pegasus_pipeline/ingesters/ingest_aif.py
@@ -2,57 +2,36 @@
 import argparse
 import logging
 from pathlib import Path
+from typing import Optional
 
 from pegasus_wrapper.key_value import ZipKeyValueStore
-from rdflib import Graph, URIRef
+from rdflib import Graph
 from tqdm import tqdm
 from vistautils.key_value import byte_key_value_sink_from_params
 from vistautils.parameters import Parameters
 
-from cdse_covid.pegasus_pipeline.merge_aif.merge_aif import get_claims
-
 logger = logging.getLogger()
 
 
-def get_doc_id(graph: Graph, claim: URIRef) -> str:
-    """Get ID of document."""
-    doc_id_result = set(
-        graph.query(
-            """
-        SELECT ?doc_id
-        WHERE {
-            ?c aida:justifiedBy ?justifiedBy .
+def get_abstract_doc_id(graph: Graph) -> Optional[str]:
+    """Get doc id from any text justification element."""
+    query = """
+    SELECT ?doc_id
+    WHERE {
+        ?e a aida:TextJustification .
 
-            ?justifiedBy aida:endOffsetInclusive ?end ;
-                aida:startOffset ?start ;
-                aida:source ?doc_id .
-        }
-        """,
-            initBindings={"c": claim[0]},
-        )
-    )
+        ?e aida:source ?doc_id .
+    }
+    """
+    doc_id = set(graph.query(query))
+    if doc_id:
+        doc_id_pair = tuple(doc_id.pop())
+        return str(doc_id_pair[0].value) + ".ttl"
 
-    # If there's no justification, there must be a sourceDocument
-    if not doc_id_result:
-        doc_id_result = set(
-            graph.query(
-                """
-            SELECT ?doc_id
-            WHERE {
-                ?c aida:sourceDocument ?doc_id
-            }
-            """,
-                initBindings={"c": claim[0]},
-            )
-        )
-        doc_id = doc_id_result.pop()
-        return str(doc_id[0].value) + ".ttl"
-
-    doc_id = doc_id_result.pop()
-    return str(doc_id[0].value) + ".ttl"
+    return None
 
 
-def save_to_key_value_store(aif_dir: Path, kvs_path: Path, *, is_uiuc: bool) -> ZipKeyValueStore:
+def save_to_key_value_store(aif_dir: Path, kvs_path: Path) -> ZipKeyValueStore:
     """Save AIF to a ZipKeyValueStore."""
     local_params = Parameters.from_mapping({"output": {"type": "zip", "path": str(kvs_path)}})
 
@@ -62,27 +41,18 @@ def save_to_key_value_store(aif_dir: Path, kvs_path: Path, *, is_uiuc: bool) -> 
         ):  # TQDM doesn't actually work b/c it's a generator
             g = Graph()
             g = g.parse(source=aif_file, format="turtle")
-            key = aif_file.name
-            if is_uiuc:
-                claims = get_claims(g)
-                doc_ids = set()
-                if not claims:
-                    continue
-                for claim in claims:
-                    doc_id = get_doc_id(g, claim)
-                    doc_ids.add(doc_id)
-                if len(doc_ids) > 1:
-                    raise RuntimeError(f"Claims from multiple documents found in {aif_file.name}")
-                else:
-                    key = doc_ids.pop()
-            sink[key] = g.serialize(format="turtle", encoding="utf-8")
+
+            key = get_abstract_doc_id(g)
+
+            if key:
+                sink[key] = g.serialize(format="turtle", encoding="utf-8")
 
     return ZipKeyValueStore(path=kvs_path)
 
 
-def main(aif_dir: Path, aif_as_zip: Path, *, is_uiuc: bool) -> None:
+def main(aif_dir: Path, aif_as_zip: Path) -> None:
     """Entrypoint to AIF ingest script."""
-    save_to_key_value_store(aif_dir, aif_as_zip, is_uiuc=is_uiuc)
+    save_to_key_value_store(aif_dir, aif_as_zip)
     logger.info("Serialized all TTL files to: %s", aif_as_zip)
 
 
@@ -90,6 +60,5 @@ if __name__ == "__main__":
     p = argparse.ArgumentParser(description=__doc__)
     p.add_argument("--aif-dir", type=Path, help="option help")
     p.add_argument("--aif-as-zip", type=Path)
-    p.add_argument("--is-uiuc", action="store_true", default=False)
     args = p.parse_args()
-    main(args.aif_dir, args.aif_as_zip, is_uiuc=args.is_uiuc)
+    main(args.aif_dir, args.aif_as_zip)
