+++ cdse_covid/pegasus_pipeline/merge/aif_models.py
@@ -1,11 +1,13 @@
 """AIF Models."""
+from typing import Optional
+
 from rdflib.term import Node
 
 
 class MergeNode(Node):
     """Node to handle merge provenance."""
 
-    def __init__(self, provenance: str = "") -> None:
+    def __init__(self, provenance: Optional[str] = "") -> None:
         """Init MergeNode."""
         super().__init__()
         self.provenance = provenance
@@ -18,7 +20,9 @@ class MergeNode(Node):
 class Span(MergeNode):
     """Node to handle Span."""
 
-    def __init__(self, start: int, end: int, source: str = "", provenance: str = "") -> None:
+    def __init__(
+        self, start: int, end: int, source: str = "", provenance: Optional[str] = ""
+    ) -> None:
         """Init Span."""
         super().__init__(provenance=provenance)
         self.start = start
